Improve ergonomics of client write error handling.
This is a farily minimal changeset, but is a breaking change.
This introduces a `RaftStorage::ShutdownError` associated type.
This allows for the Raft system to differentiate between fatal storage
errors which should cause the system to shutdown vs errors which should
be propagated back to the client for application specific error handling.
These changes only apply to the
`RaftStorage::apply_entry_to_state_machine` method.

Overall, this constitutes a nice improvement to the RaftStorage
interface, as a flat result type may be used for propagating errors. No
result nesting required.